+++ SourceryRuntime/Sources/TypeName.swift
@@ -30,7 +30,7 @@ public protocol Typed {
 }
 
 /// Describes name of the type used in typed declaration (variable, method parameter or return value etc.)
-public final class TypeName: NSObject, AutoCoding, AutoEquatable, AutoDiffable, AutoJSExport, LosslessStringConvertible {
+@objcMembers public final class TypeName: NSObject, AutoCoding, AutoEquatable, AutoDiffable, AutoJSExport, LosslessStringConvertible {
 
     /// :nodoc:
     public init(_ name: String,
@@ -231,7 +231,7 @@ public final class TypeName: NSObject, AutoCoding, AutoEquatable, AutoDiffable, 
 }
 
 /// Descibes Swift generic type parameter
-public final class GenericTypeParameter: NSObject, SourceryModel {
+@objcMembers public final class GenericTypeParameter: NSObject, SourceryModel {
 
     /// Generic parameter type name
     public let typeName: TypeName
@@ -263,7 +263,7 @@ public final class GenericTypeParameter: NSObject, SourceryModel {
 }
 
 /// Descibes Swift generic type
-public final class GenericType: NSObject, SourceryModel {
+@objcMembers public final class GenericType: NSObject, SourceryModel {
     /// the name of the base type.
     ///
     /// `Array<Int>`'s GenericType.name is `Array`
@@ -295,7 +295,7 @@ public final class GenericType: NSObject, SourceryModel {
 }
 
 /// Describes tuple type element
-public final class TupleElement: NSObject, SourceryModel, Typed {
+@objcMembers public final class TupleElement: NSObject, SourceryModel, Typed {
 
     /// Tuple element name
     public let name: String
@@ -332,7 +332,7 @@ public final class TupleElement: NSObject, SourceryModel, Typed {
 }
 
 /// Describes tuple type
-public final class TupleType: NSObject, SourceryModel {
+@objcMembers public final class TupleType: NSObject, SourceryModel {
 
     /// Type name used in declaration
     public let name: String
@@ -362,7 +362,7 @@ public final class TupleType: NSObject, SourceryModel {
 }
 
 /// Describes array type
-public final class ArrayType: NSObject, SourceryModel {
+@objcMembers public final class ArrayType: NSObject, SourceryModel {
 
     /// Type name used in declaration
     public let name: String
@@ -399,7 +399,7 @@ public final class ArrayType: NSObject, SourceryModel {
 }
 
 /// Describes dictionary type
-public final class DictionaryType: NSObject, SourceryModel {
+@objcMembers public final class DictionaryType: NSObject, SourceryModel {
 
     /// Type name used in declaration
     public let name: String
@@ -449,7 +449,7 @@ public final class DictionaryType: NSObject, SourceryModel {
 }
 
 /// Describes closure type
-public final class ClosureType: NSObject, SourceryModel {
+@objcMembers public final class ClosureType: NSObject, SourceryModel {
 
     /// Type name used in declaration with stripped whitespaces and new lines
     public let name: String